+++ truffle/src/com.oracle.truffle.api.dsl/src/com/oracle/truffle/api/dsl/Specialization.java
@@ -29,6 +29,8 @@ import com.oracle.truffle.api.frame.Frame;
 import com.oracle.truffle.api.frame.MaterializedFrame;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.Node;
+import com.oracle.truffle.api.nodes.UnexpectedResultException;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -140,9 +142,12 @@ public @interface Specialization {
      * <p>
      * If an event guard exception is triggered then all instantiations of this specialization are
      * removed. If one of theses exceptions is thrown once then no further instantiations of this
-     * specialization are going to be created for this node. A specialization that rewrites on an
-     * exception must ensure that no non-repeatable side-effect is caused until the rewrite is
-     * triggered.
+     * specialization are going to be created for this node.
+     *
+     * In case of explicitly declared {@link UnexpectedResultException}s, the result from the
+     * exception will be used. For all other exception types, the next available specialization will
+     * be executed, so that the original specialization must ensure that no non-repeatable
+     * side-effect is caused until the rewrite is triggered.
      * </p>
      *
      * <b>Example usage:</b>